+++ Lenden.Backend/Lenden.Core/ExpenseFeatures/ExpenseSplitEntity.cs
@@ -9,15 +9,15 @@ public class ExpenseSplitEntity
         
     }
 
-    public ExpenseSplitEntity(int  expenseId, int splitterId, decimal amount)
+    public ExpenseSplitEntity(long  expenseId, int splitterId, decimal amount)
     {
         SplitterId = splitterId;
         Amount = amount;
         ExpenseId = expenseId;
     }
     
-    public int Id { get; set; }
-    public int ExpenseId { get; set; }
+    public long Id { get; set; }
+    public long ExpenseId { get; set; }
     public virtual ExpenseEntity Expense { get; set; }
     public int SplitterId { get; set; }
     public virtual UserEntity Splitter { get; set; }
