+++ Lenden.Backend/Lenden.Data/Configurations/ExpenseSplitEntityConfiguration.cs
@@ -28,7 +28,7 @@ public class ExpenseSplitEntityConfiguration: IEntityTypeConfiguration<ExpenseSp
         builder.Property(e=>e.ExpenseId)
             .HasColumnName("expense_id")
             .IsRequired()
-            .HasColumnType("integer");
+            .HasColumnType("bigint");
         
         builder.HasOne(ep => ep.Expense)
             .WithMany()
